ClickHandler removed, avoid call loop duirng long press on a webview between BrowserActivity and LightningView
