+++ app/src/main/java/acr/browser/lightning/bus/BrowserEvents.java
@@ -1,6 +1,7 @@
 package acr.browser.lightning.bus;
 
 import android.support.annotation.IdRes;
+import android.support.annotation.Nullable;
 import android.support.annotation.StringRes;
 
 import acr.browser.lightning.R;
